Split app config options to a sub config
