+++ examples/node/hello-sdl-tcp/index.js
@@ -43,7 +43,7 @@ class AppClient {
             .setType(SDL.rpc.enums.FileType.GRAPHIC_PNG)
             .setPersistent(true);
 
-        this._appConfig = new SDL.manager.AppConfig()
+        this._lifecycleConfig = new SDL.manager.LifecycleConfig()
             .setAppId(CONFIG.appId)
             .setAppName(CONFIG.appName)
             .setLanguageDesired(SDL.rpc.enums.Language.EN_US)
@@ -53,6 +53,9 @@ class AppClient {
             ])
             .setTransportConfig(new SDL.transport.TcpClientConfig(CONFIG.host, CONFIG.port))
             .setAppIcon(file);
+
+        this._appConfig = new SDL.manager.AppConfig()
+            .setLifecycleConfig(this._lifecycleConfig);
 
         const managerListener = new SDL.manager.SdlManagerListener();
         managerListener
