+++ examples/webengine/hello-sdl/index.html
@@ -47,11 +47,14 @@
                     .setType(SDL.rpc.enums.FileType.GRAPHIC_PNG)
                     .setPersistent(true);
 
-                this._appConfig = new SDL.manager.AppConfig()
+                this._lifecycleConfig = new SDL.manager.LifecycleConfig()
                     .loadManifest(sdl_manifest)
                     .setLanguageDesired(SDL.rpc.enums.Language.EN_US)
                     .setHmiDisplayLanguageDesired(SDL.rpc.enums.Language.EN_US)
                     .setTransportConfig(new SDL.transport.WebSocketClientConfig());
+
+                this._appConfig = new SDL.manager.AppConfig()
+                    .setLifecycleConfig(this._lifecycleConfig);
 
                 const managerListener = new SDL.manager.SdlManagerListener();
                 managerListener
