+++ lib/js/src/manager/AppConfig.js
@@ -33,6 +33,7 @@
 import { Language } from '../rpc/enums/Language.js';
 import { AppHMIType } from '../rpc/enums/AppHMIType.js';
 import { Version } from '../util/Version.js';
+import { LifecycleConfig } from './LifecycleConfig.js';
 
 class AppConfig {
     /**
@@ -40,334 +41,25 @@ class AppConfig {
      * @class
      */
     constructor () {
-        this._transportConfig = null;
-        this._appId = null; // required
-        this._appName = null; // required
-        this._appIcon = null;
-        this._shortAppName = null;
-        this._ttsName = null;
-        this._vrSynonyms = null;
-        this._isMediaApp = false;
-        this._languageDesired = Language.EN_US;
-        this._hmiDisplayLanguageDesired = Language.EN_US;
-        this._appTypes = [AppHMIType.DEFAULT]; // hmiTypes
-        this._dayColorScheme = null;
-        this._nightColorScheme = null;
-        this._minimumRpcVersion = new Version(1, 0, 0);
-        this._minimumProtocolVersion = new Version(1, 0, 0);
+        this._lifecycleConfig = null;
     }
 
     /**
-     * Set the TransportConfig instance.
-     * @param {_TransportConfigBase} transportConfig - An instance of TransportConfig.
+     * Set the LifecycleConfig instance.
+     * @param {LifecycleConfig} lifecycleConfig - An instance of LifecycleConfig.
      * @returns {AppConfig} - A reference to this instance to support method chaining.
      */
-    setTransportConfig (transportConfig) {
-        this._transportConfig = transportConfig;
+    setLifecycleConfig (lifecycleConfig) {
+        this._lifecycleConfig = lifecycleConfig;
         return this;
     }
 
     /**
-     * Get the TransportConfig instance.
-     * @returns {_TransportConfigBase} - The TransportConfig.
+     * Get the LifecycleConfig instance.
+     * @returns {LifecycleConfig} - The LifecycleConfig.
      */
-    getTransportConfig () {
-        return this._transportConfig;
-    }
-
-    /**
-     * Set the App ID. Use a full App Id.
-     * @param {String} appId - The full App ID.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setAppId (appId) {
-        this._appId = appId;
-        return this;
-    }
-
-    /**
-     * Get the App ID.
-     * @returns {String} - The full app ID.
-     */
-    getAppId () {
-        return this._appId;
-    }
-
-    /**
-     * Set the app name.
-     * @param {String} appName - The app name.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setAppName (appName) {
-        this._appName = appName;
-        return this;
-    }
-
-    /**
-     * Get the app name.
-     * @returns {String} - The app name.
-     */
-    getAppName () {
-        return this._appName;
-    }
-
-    /**
-     * Set the app icon.
-     * @param {SdlArtwork} sdlArtwork - An instance of SdlArtwork.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setAppIcon (sdlArtwork) {
-        this._appIcon = sdlArtwork;
-        return this;
-    }
-
-    /**
-     * Get the app icon.
-     * @returns {SdlArtwork} - An instance of SdlArtwork.
-     */
-    getAppIcon () {
-        return this._appIcon;
-    }
-
-    /**
-     * Set the short app name.
-     * @param {String} shortAppName - The short app name.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setShortAppName (shortAppName) {
-        this._shortAppName = shortAppName;
-        return this;
-    }
-
-    /**
-     * Get the short app name.
-     * @returns {String} - The short app name.
-     */
-    getShortAppName () {
-        return this._shortAppName;
-    }
-
-    /**
-     * Set the TTS Names.
-     * @param {Array<TTSChunk>} ttsName - An array of TTSChunk structs.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setTtsName (ttsName) {
-        this._ttsName = ttsName;
-        return this;
-    }
-
-    /**
-     * Get the TTS names.
-     * @returns {Array<TTSChunk>} - An array of TTSChunk structs.
-     */
-    getTtsName () {
-        return this._ttsName;
-    }
-
-    /**
-     * Set the VR Synonyms.
-     * @param {Array<String>} vrSynonyms - An array of strings representing VR Synonyms.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setVrSynonyms (vrSynonyms) {
-        this._vrSynonyms = vrSynonyms;
-        return this;
-    }
-
-    /**
-     * Get the VR Synonyms.
-     * @returns {Array<String>} - An array of strings representing VR Synonyms.
-     */
-    getVrSynonyms () {
-        return this._vrSynonyms;
-    }
-
-    /**
-     * Set the desired language of the application.
-     * @param {Language} languageDesired - A Language enum value.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setLanguageDesired (languageDesired) {
-        this._languageDesired = languageDesired;
-        return this;
-    }
-
-    /**
-     * Get the desired langauge of the application.
-     * @returns {Language} - A Language enum value.
-     */
-    getLanguageDesired () {
-        return this._languageDesired;
-    }
-
-    /**
-     * Set the desired HMI Display Language.
-     * @param {Language} hmiDisplayLanguageDesired - A Language enum value.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setHmiDisplayLanguageDesired (hmiDisplayLanguageDesired) {
-        this._hmiDisplayLanguageDesired = hmiDisplayLanguageDesired;
-        return this;
-    }
-
-    /**
-     * Get the desired HMI Display Language.
-     * @returns {Language} - A Langauge enum value.
-     */
-    getHmiDisplayLanguageDesired () {
-        return this._hmiDisplayLanguageDesired;
-    }
-
-    /**
-     * Set the application's types.
-     * @param {Array<AppHMIType>} appTypes - An array of ordered AppHMIType enum values.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setAppTypes (appTypes) {
-        this._appTypes = appTypes;
-        if (appTypes !== null && appTypes !== undefined) {
-            this._isMediaApp = appTypes.includes(AppHMIType.MEDIA);
-        }
-        return this;
-    }
-
-    /**
-     * Get the application's types.
-     * @returns {Array<AppHMIType>} - An array of ordered AppHMIType enum values.
-     */
-    getAppTypes () {
-        return this._appTypes;
-    }
-
-
-    /**
-     * Set the application's day color scheme.
-     * @param {TemplateColorScheme} dayColorScheme - A TemplateColorScheme instance.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setDayColorScheme (dayColorScheme) {
-        this._dayColorScheme = dayColorScheme;
-        return this;
-    }
-
-    /**
-     * Get the application's day color scheme.
-     * @returns {TemplateColorScheme} - A TemplateColorScheme instance.
-     */
-    getDayColorScheme () {
-        return this._dayColorScheme;
-    }
-
-    /**
-     * Set the application's night color scheme.
-     * @param {TemplateColorScheme} nightColorScheme - A TemplateColorScheme instance.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setNightColorScheme (nightColorScheme) {
-        this._nightColorScheme = nightColorScheme;
-        return this;
-    }
-
-    /**
-     * Get the application's night color scheme.
-     * @returns {TemplateColorScheme} - A TemplateColorScheme instance.
-     */
-    getNightColorScheme () {
-        return this._nightColorScheme;
-    }
-
-    /**
-     * Set the minimum RPC version that will be permitted to connect. If the RPC version of the head unit connected is below this version, an UnregisterAppInterface will be sent.
-     * @param {Version} minimumRpcVersion - A Version instance.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setMinimumRpcVersion (minimumRpcVersion) {
-        this._minimumRpcVersion = minimumRpcVersion;
-        return this;
-    }
-
-    /**
-     * Get the minimum RPC version.
-     * @returns {Version} - A Version instance.
-     */
-    getMinimumRpcVersion () {
-        return this._minimumRpcVersion;
-    }
-
-
-    /**
-     * Sets the minimum protocol version that will be permitted to connect. If the protocol version of the head unit connected is below this version, the app will disconnect with an EndService protocol message and will not register.
-     * @param {Version} minimumProtocolVersion - A Version instance.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    setMinimumProtocolVersion (minimumProtocolVersion) {
-        this._minimumProtocolVersion = minimumProtocolVersion;
-        return this;
-    }
-
-    /**
-     * Get the minimum protocol version supported by the app.
-     * @returns {Version} - A Version instance.
-     */
-    getMinimumProtocolVersion () {
-        return this._minimumProtocolVersion;
-    }
-
-    /**
-     * Load manifest file of embedded Web Application
-     * @param {Object} manifest - A manifest JSON object.
-     * @returns {AppConfig} - A reference to this instance to support method chaining.
-     */
-    loadManifest (manifest) {
-        if (typeof manifest !== 'object') {
-            throw new Error('Wrong parameter, manifest must be object');
-        }
-        const manifestKeys = Object.keys(manifest);
-        const missingKeys = [
-            'appId',
-            'appName',
-            'category',
-            'minProtocolVersion',
-            'minRpcVersion',
-        ].filter(key => !manifestKeys.includes(key));
-        if (missingKeys.length > 0) {
-            throw new Error(`Required parameters are missing in the manifest: ${missingKeys.join(', ')}`);
-        }
-        manifestKeys.forEach((key) => {
-            switch (key) {
-                case 'appId':
-                    this.setAppId(manifest[key]);
-                    break;
-                case 'appName':
-                    this.setAppName(manifest[key]);
-                    break;
-                case 'category':
-                case 'additionalCategories': {
-                    // merge the category and additionalCategories into one de-duplicated array
-                    const categories = (Array.isArray(manifest[key]) ? manifest[key] : [manifest[key]])
-                        .map((elem) => {
-                            const key = AppHMIType.keyForValue(elem);
-                            return key ? AppHMIType[key] : '';
-                        })
-                        .concat(this.getAppTypes() || [])
-                        .filter((elem, pos, arr) => elem && (arr.indexOf(elem) === pos));
-                    this.setAppTypes(categories);
-                    break;
-                }
-                case 'minProtocolVersion': {
-                    const version = new Version().fromString(manifest[key]);
-                    this.setMinimumProtocolVersion(version);
-                    break;
-                }
-                case 'minRpcVersion': {
-                    const version = new Version().fromString(manifest[key]);
-                    this.setMinimumRpcVersion(version);
-                    break;
-                }
-            }
-        });
-        return this;
+    getLifecycleConfig () {
+        return this._lifecycleConfig;
     }
 }
