+++ lib/js/src/manager/SdlManager.js
@@ -56,6 +56,7 @@ class SdlManager extends _SdlManagerBase {
         this._permissionManager = null;
         this._fileManager = null;
         this._screenManager = null;
+        this._lifecycleConfig = null;
 
         // NOTE: use "self" for calling "this" inside async methods until resolved:
         // https://github.com/rpetrich/babel-plugin-transform-async-to-promises/issues/54
@@ -174,23 +175,23 @@ class SdlManager extends _SdlManagerBase {
      */
     async _onReady () {
         // Set the app icon
-        if (this._appConfig !== null && this._appConfig.getAppIcon() !== null) {
+        if (this._lifecycleConfig !== null && this._lifecycleConfig.getAppIcon() !== null) {
             if (
                 this._fileManager !== null
                 && this._fileManager._getState() === _SubManagerBase.READY
-                && !this._fileManager.hasUploadedFile(this._appConfig.getAppIcon())
+                && !this._fileManager.hasUploadedFile(this._lifecycleConfig.getAppIcon())
             ) {
                 // Icon not uploaded yet
-                const uploadResult = await this._fileManager.uploadArtwork(this._appConfig.getAppIcon());
+                const uploadResult = await this._fileManager.uploadArtwork(this._lifecycleConfig.getAppIcon());
                 if (uploadResult) {
                     await this.sendRpc(
-                        new SetAppIcon().setFileName(this._appConfig.getAppIcon().getName())
+                        new SetAppIcon().setFileName(this._lifecycleConfig.getAppIcon().getName())
                     );
                 }
             } else {
                 // Icon already uploaded
                 await this.sendRpc(
-                    new SetAppIcon().setFileName(this._appConfig.getAppIcon().getName())
+                    new SetAppIcon().setFileName(this._lifecycleConfig.getAppIcon().getName())
                 );
             }
         }
@@ -222,7 +223,7 @@ class SdlManager extends _SdlManagerBase {
     _checkLifecycleConfiguration () {
         const actualLanguage = this._lifecycleManager.getRegisterAppInterfaceResponse().getLanguage();
 
-        if (actualLanguage !== null && actualLanguage !== this._appConfig.getHmiDisplayLanguageDesired()) {
+        if (actualLanguage !== null && actualLanguage !== this._lifecycleConfig.getHmiDisplayLanguageDesired()) {
             // HMI language doesn't match the app's desired display language
             const lifecycleConfigUpdate = this._managerListener.managerShouldUpdateLifecycle(actualLanguage);
 
@@ -238,16 +239,16 @@ class SdlManager extends _SdlManagerBase {
                 this.sendRpc(changeRegistration)
                     .then((response) => {
                         if (lifecycleConfigUpdate.getAppName() !== null) {
-                            this._appConfig.setAppName(lifecycleConfigUpdate.getAppName());
+                            this._lifecycleConfig.setAppName(lifecycleConfigUpdate.getAppName());
                         }
                         if (lifecycleConfigUpdate.getShortAppName() !== null) {
-                            this._appConfig.setShortAppName(lifecycleConfigUpdate.getShortAppName());
+                            this._lifecycleConfig.setShortAppName(lifecycleConfigUpdate.getShortAppName());
                         }
                         if (lifecycleConfigUpdate.getTtsName() !== null) {
-                            this._appConfig.setTtsName(lifecycleConfigUpdate.getTtsName());
+                            this._lifecycleConfig.setTtsName(lifecycleConfigUpdate.getTtsName());
                         }
                         if (lifecycleConfigUpdate.getVoiceRecognitionCommandNames() !== null) {
-                            this._appConfig.setVrSynonyms(lifecycleConfigUpdate.getVoiceRecognitionCommandNames());
+                            this._lifecycleConfig.setVrSynonyms(lifecycleConfigUpdate.getVoiceRecognitionCommandNames());
                         }
                     });
             }
@@ -314,23 +315,29 @@ class SdlManager extends _SdlManagerBase {
     start () {
         if (this._lifecycleManager === null) {
             if (this.getAppConfig() !== null) {
-                if (this.getAppConfig().getTransportConfig() === null) {
-                    throw new Error('AppConfig must have a transport config defined');
+                this._lifecycleConfig = this.getAppConfig().getLifecycleConfig();
+
+                if (this._lifecycleConfig === null) {
+                    throw new Error('AppConfig must have a LifecycleConfig defined');
                 }
 
-                if (this.getAppConfig().getAppName() === null) {
-                    throw new Error('AppConfig must have an app name defined');
+                if (this._lifecycleConfig.getTransportConfig() === null) {
+                    throw new Error('LifecycleConfig must have a transport config defined');
                 }
 
-                if (this.getAppConfig().getAppId() === null) {
-                    throw new Error('AppConfig must have an app ID defined');
+                if (this._lifecycleConfig.getAppName() === null) {
+                    throw new Error('LifecycleConfig must have an app name defined');
+                }
+
+                if (this._lifecycleConfig.getAppId() === null) {
+                    throw new Error('LifecycleConfig must have an app ID defined');
                 }
 
                 if (this.getManagerListener() === null) {
                     throw new Error('An SdlManagerListener must be defined');
                 }
 
-                this._lifecycleManager = new _LifecycleManager(this.getAppConfig(), this._lifecycleListener);
+                this._lifecycleManager = new _LifecycleManager(this._lifecycleConfig, this._lifecycleListener);
 
                 /* FIXME: setSdlSecurity and related methods/classes do not exist
                 if (this._sdlSecList.length > 0) {
